+++ pyirena/gui/_qt.py
@@ -114,10 +114,12 @@ QDoubleValidator = QtGui.QDoubleValidator
 QFont = QtGui.QFont
 QIcon = QtGui.QIcon
 QIntValidator = QtGui.QIntValidator
+QKeySequence = QtGui.QKeySequence
 QPainterPath = QtGui.QPainterPath
 QPainterPathStroker = QtGui.QPainterPathStroker
 QPen = QtGui.QPen
 QPixmap = QtGui.QPixmap
+QShortcut = QtGui.QShortcut
 QTransform = QtGui.QTransform
 
 __all__ = [
@@ -138,6 +140,7 @@ __all__ = [
     "QRectF", "QThread", "QTimer", "QUrl",
     # QtGui
     "QAction", "QBrush", "QCloseEvent", "QColor", "QDesktopServices",
-    "QDoubleValidator", "QFont", "QIcon", "QIntValidator", "QPainterPath",
-    "QPainterPathStroker", "QPen", "QPixmap", "QTransform",
+    "QDoubleValidator", "QFont", "QIcon", "QIntValidator", "QKeySequence",
+    "QPainterPath", "QPainterPathStroker", "QPen", "QPixmap", "QShortcut",
+    "QTransform",
 ]